# go-server/openapi.yaml
openapi: 3.0.0
...
info:
  title: bear-trak-backend
  version: "1.0"
  description: This API services the iOS app BearTrak. It provides all necessary backend data.
  contact:
    name: Benjamin Koppe
    url: github.com/benkoppe
    email: koppe.development@gmail.com
  license:
    name: GPLv3
    url: "https://www.gnu.org/licenses/gpl-3.0.en.html"
servers:
  - url: "https://beartrakapi.thekoppe.com"
    description: Service deployment.
  - url: "http://localhost:3000"
    description: Localhost deployment.
paths:
  /v1/alerts:
    get:
      summary: Alerts
      tags:
        - beartrak
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1Alerts
      x-stoplight:
        id: i81d677y6pf4u
      description: Returns all of BearTrak's active alerts.
  /v1/dining:
    get:
      summary: Dining
      tags:
        - dining
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Eatery"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1Dining
      x-stoplight:
        id: mr1vidtnkgibv
      description: Returns all necessary data for BearTrak's dining section.
  /v1/gyms:
    get:
      summary: Gyms
      tags:
        - gyms
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gym"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1Gyms
      x-stoplight:
        id: bgkkzgl4exihm
      description: Returns all necessary data for BearTrak's gym section.
  /v1/gyms/capacities:
    get:
      summary: Gym Capacities
      tags:
        - gyms
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GymCapacityData"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1GymCapacities
      x-stoplight:
        id: a8uzxemghpcxz
      description: Returns all time-logged gym capacity data
  /v1/gyms/capacities/predictions:
    get:
      summary: Gym Capacity Predictions
      tags:
        - gyms
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GymCapacityPredictions"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1GymCapacityPredictions
      x-stoplight:
        id: aglc6o7iqsbqo
      description: Returns all time-logged gym capacity prediction data
  /v1/study:
    get:
      summary: Study
      tags:
        - study
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyData"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1Study
      x-stoplight:
        id: 4ahnm07h1tk19
      description: Returns all necessary data for BearTrak's study section.
    parameters: []
  /v1/transit/routes:
    get:
      summary: Routes
      tags:
        - transit
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusRoute"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1TransitRoutes
      x-stoplight:
        id: x9u9nrkadrz7r
      description: "Returns non time-sensitive, route-related data for BearTrak's transit section."
  /v1/transit/vehicles:
    get:
      summary: Vehicles
      tags:
        - transit
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1TransitVehicles
      x-stoplight:
        id: skii58kqavo15
      description: "Returns time-sensitive, vehicle-related data for BearTrak's transit section."
  /v1/events:
    get:
      summary: Events
      tags:
        - events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1Events
      x-stoplight:
        id: ohky43ld8kd45
      description: Returns event data for BearTrak's events section.
  /v1/dining/user:
    parameters: []
    get:
      summary: Get Information
      tags: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiningUser"
        "401":
          description: Unauthorized
        default:
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1DiningUser
      x-stoplight:
        id: oypusjbwr7hrc
      parameters:
        - $ref: "#/components/parameters/DiningUserSession"
    post:
      summary: Register
      tags:
        - dining
        - diningUser
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - $ref: "#/components/schemas/DiningUser"
        "400":
          description: "Bad Request: device already exists"
          content: {}
        "401":
          description: "Unauthorized: bad sessionId"
          content: {}
        default:
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: postV1DiningUser
      x-stoplight:
        id: bk7qsagm0f0yz
      description: Registers a new user given a device and session
      parameters:
        - $ref: "#/components/parameters/DiningUserSession"
        - $ref: "#/components/parameters/DiningUserDeviceId"
        - $ref: "#/components/parameters/DiningUserDevicePin"
    delete:
      summary: Delete
      tags:
        - dining
        - diningUser
      responses:
        "200":
          description: OK deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          description: Unauthorized
          content: {}
        default:
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: deleteV1DiningUser
      x-stoplight:
        id: lcyrjqsgno3u9
      description: Deletes a user given a session
      parameters:
        - $ref: "#/components/parameters/DiningUserSession"
  /v1/dining/user/session:
    get:
      summary: Refresh Token
      tags:
        - dining
        - diningUser
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content: {}
        default:
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1DiningUserSession
      x-stoplight:
        id: kqdhzz2k3y749
      description: Refreshes a session given a user device
      parameters:
        - $ref: "#/components/parameters/DiningUserDeviceId"
        - $ref: "#/components/parameters/DiningUserDevicePin"
    parameters: []
  /v1/dining/user/accounts:
    get:
      summary: Dining Accounts
      tags:
        - dining
        - diningUser
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiningUserAccount"
        "401":
          description: Unauthorized
          content: {}
        default:
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getV1DiningUserAccounts
      x-stoplight:
        id: o3dflz1hb8swp
      description: Returns a dining user's transaction accounts given a session
      parameters:
        - $ref: "#/components/parameters/DiningUserSession"
    parameters: []
components:
  schemas:
    Alert:
      title: Alert
      x-stoplight:
        id: ae6llgzegnzv5
      type: object
      x-tags:
        - beartrak
      required:
        - id
        - title
        - message
        - enabled
        - showOnce
        - maxBuild
        - button
      properties:
        id:
          type: integer
          x-stoplight:
            id: 80tpwhx3zfpi6
        title:
          type: string
          x-stoplight:
            id: yem4lvu39oask
        message:
          type: string
          x-stoplight:
            id: bp5ir995yemxs
        enabled:
          type: boolean
          x-stoplight:
            id: 5cgo7ogy7oizn
        showOnce:
          type: boolean
          x-stoplight:
            id: asz4nk5bkpxut
        maxBuild:
          type: integer
          x-stoplight:
            id: irt2vgltn41jt
          nullable: true
        button:
          type: object
          x-stoplight:
            id: 149c9ieot39qz
          required:
            - title
            - url
          nullable: true
          properties:
            title:
              type: string
              x-stoplight:
                id: 897j0kv2lom6r
            url:
              type: string
              x-stoplight:
                id: 9rhio1ewe055f
              format: uri
        minMinutesSinceDownload:
          type: integer
          x-stoplight:
            id: meq2g0rg3k557
    Error:
      title: Error
      x-stoplight:
        id: csh4s0fu2c0dz
      type: object
      x-tags:
        - beartrak
      required:
        - code
        - message
      properties:
        code:
          type: integer
          x-stoplight:
            id: 3yyr0yzzvfru1
        message:
          type: string
          x-stoplight:
            id: 3kla3y1y92kn5
...
    Eatery:
      title: Eatery
      x-stoplight:
        id: t3d6ig3lqsua5
      type: object
      x-tags:
        - dining
      required:
        - id
        - name
        - nameShort
        - imagePath
        - latitude
        - longitude
        - location
        - hours
        - region
        - payMethods
        - categories
        - nextWeekEvents
        - allWeekMenu
      properties:
        id:
          type: integer
          x-stoplight:
            id: 18a2zedwfskoc
        name:
          type: string
          x-stoplight:
            id: qxk0k8jknfhwf
        nameShort:
          type: string
          x-stoplight:
            id: dswz6jvcu0njw
        imagePath:
          type: string
          x-stoplight:
            id: 0lv5n6id9dgjr
        latitude:
          type: number
          x-stoplight:
            id: dst3snzwlgj8f
          format: double
        longitude:
          type: number
          x-stoplight:
            id: aq7jp1pchfk33
          format: double
        location:
          type: string
          x-stoplight:
            id: ijnjtu0smflyv
        hours:
          x-stoplight:
            id: w9vicfj2xyci2
          type: array
          items:
            $ref: "#/components/schemas/Hours"
        region:
          x-stoplight:
            id: g762nv65cuibu
          type: string
          example: central
        payMethods:
          type: array
          x-stoplight:
            id: nlpfgxpn5obeb
          items:
            x-stoplight:
              id: tj2iccdcf7xpp
            type: string
            example: swipes
        categories:
          type: array
          x-stoplight:
            id: aizzbsbp6idt6
          items:
            x-stoplight:
              id: mcasfw3bcvr08
            type: string
            enum:
              - convenienceStore
              - cafe
              - diningRoom
              - coffeeShop
              - cart
              - foodCourt
            example: diningRoom
        nextWeekEvents:
          type: object
          x-stoplight:
            id: 2k2uouqvuivog
          required:
            - monday
            - tuesday
            - wednesday
            - thursday
            - friday
            - saturday
            - sunday
          properties:
            monday:
              x-stoplight:
                id: rxtsxv5xozp0t
              type: array
              nullable: true
              items:
                $ref: "#/components/schemas/EateryEvent"
            tuesday:
              x-stoplight:
                id: 1qayemfv8za3l
              type: array
              nullable: true
              items:
                $ref: "#/components/schemas/EateryEvent"
            wednesday:
              type: array
              x-stoplight:
                id: tby1a09yx1hdd
              nullable: true
              items:
                $ref: "#/components/schemas/EateryEvent"
            thursday:
              type: array
              x-stoplight:
                id: m3d80t4db5273
              nullable: true
              items:
                $ref: "#/components/schemas/EateryEvent"
            friday:
              type: array
              x-stoplight:
                id: wjb3fxmxxez8t
              nullable: true
              items:
                $ref: "#/components/schemas/EateryEvent"
            saturday:
              type: array
              x-stoplight:
                id: 1zcuyrewwph08
              nullable: true
              items:
                $ref: "#/components/schemas/EateryEvent"
            sunday:
              type: array
              x-stoplight:
                id: uri1v7i0uw41m
              nullable: true
              items:
                $ref: "#/components/schemas/EateryEvent"
        allWeekMenu:
          type: array
          x-stoplight:
            id: nywsrv2ysf0x0
          nullable: true
          items:
            $ref: "#/components/schemas/EateryMenuCategory"
...
    Gym:
      title: Gym
      x-stoplight:
        id: arsgfiyj5hnnw
      type: object
      x-tags:
        - gyms
      required:
        - id
        - name
        - imagePath
        - latitude
        - longitude
        - hours
        - facilities
        - equipmentCategories
        - capacity
      properties:
        id:
          type: integer
          x-stoplight:
            id: e38t09tjxyc9p
        name:
          type: string
          x-stoplight:
            id: lvi3q8g6d7ygx
        imagePath:
          type: string
          x-stoplight:
            id: izei9phdx10ou
        latitude:
          type: number
          x-stoplight:
            id: o3lir1bffz9tn
          format: double
        longitude:
          type: number
          x-stoplight:
            id: yvuz0dlbv0dzt
          format: double
        hours:
          x-stoplight:
            id: iodtmc4p0kero
          type: array
          items:
            $ref: "#/components/schemas/Hours"
        facilities:
          type: array
          x-stoplight:
            id: pt2580qay9a2b
          items:
            x-stoplight:
              id: k5e4bwpntu2ou
            type: object
            required:
              - facilityType
              - name
            properties:
              facilityType:
                x-stoplight:
                  id: 1bnyqrqq6yo61
                type: string
                enum:
                  - pool
                  - basketball
                  - bowling
                  - unknown
                example: pool
              name:
                type: string
                x-stoplight:
                  id: czz0r1ru38i75
        equipmentCategories:
          type: array
          x-stoplight:
            id: oxjpkliaw0tvt
          items:
            x-stoplight:
              id: psolwrpj5sist
            type: object
            required:
              - categoryType
              - items
            properties:
              categoryType:
                x-stoplight:
                  id: vw3x0rfaemk4w
                type: string
                enum:
                  - treadmills
                  - ellipticals
                  - rowing
                  - bike
                  - lifting
                  - machines
                  - freeWeights
                  - misc
                example: treadmills
              items:
                type: array
                x-stoplight:
                  id: lf47giucu09l6
                items:
                  x-stoplight:
                    id: goz36vqbr6edz
                  type: string
        capacity:
          type: object
          x-stoplight:
            id: 8wihfhwf8dwhj
          required:
            - total
            - percentage
            - lastUpdated
          nullable: true
          properties:
            total:
              type: integer
              x-stoplight:
                id: zsa4f78a85o37
            percentage:
              type: integer
              x-stoplight:
                id: lev4eqa2vx5td
              nullable: true
            lastUpdated:
              type: string
              x-stoplight:
                id: n9pqcuhk7s66i
              format: date-time
...